+++ 03-loan/LoanManagementApp/LoanInputHandler.cs
@@ -3,19 +3,19 @@
 public class LoanInputHandler
 {
     private string[] args;
-    private const double MINIMUM_LOAN_AMOUNT = 50000;
-    private const double MINIMUM_DURATION_IN_MONTHS = 9*12;
-    private const double MAXIMUM_DURATION_IN_MONTHS = 25*12;
-    private const double MAXIMUM_PERCENTAGE = 100;
+    private const decimal MINIMUM_LOAN_AMOUNT = 50000;
+    private const decimal MINIMUM_DURATION_IN_MONTHS = 9*12;
+    private const decimal MAXIMUM_DURATION_IN_MONTHS = 25*12;
+    private const decimal MAXIMUM_PERCENTAGE = 100;
 
     public LoanInputHandler(string[] args)
     {
         this.args = args;
     }
 
-    public double GetPrincipal()
+    public decimal GetPrincipal()
     {
-        double principal = GetUserInput(0);
+        decimal principal = GetUserInput(0);
         
         if (principal < MINIMUM_LOAN_AMOUNT)
         {
@@ -37,9 +37,9 @@ public class LoanInputHandler
         return termInMonths;
     }
 
-    public double GetAnnualInterestRate()
+    public decimal GetAnnualInterestRate()
     {
-        double annualInterestRate = GetUserInput(2);
+        decimal annualInterestRate = GetUserInput(2);
 
         if (annualInterestRate <= 0 || annualInterestRate > MAXIMUM_PERCENTAGE)
         {
@@ -49,7 +49,7 @@ public class LoanInputHandler
         return annualInterestRate;
     }
 
-    private double GetUserInput(int indexArg)
+    private decimal GetUserInput(int indexArg)
     {
         string input = args[indexArg];
 
@@ -59,8 +59,8 @@ public class LoanInputHandler
             throw new ArgumentNullException();
         }
         
-        double value;
-        if (!double.TryParse(input, out value))
+        decimal value;
+        if (!decimal.TryParse(input, out value))
         {
             throw new FormatException();
         }
